# CurrencyConverter/src/main/resources/api.yaml
openapi: 3.0.0
info:
  description: API for currency conversion
  version: "1.0.0-oas3"
  title: Currency converter application
tags:
  - name: Currency Converter Api
    description: Simulator application for a toy robot
paths:
  /api/currency/convert:
    get:
      tags:
        - Currency Converter
      summary: >-
        get ammount (in source-currency), source-currency, target-currency and
        convert in to amount in target currency.
      operationId: currencyConvert
      parameters:
        - in: query
          name: source
          schema:
            type: string
        - in: query
          name: target
          schema:
            type: string
        - in: query
          name: amount
          schema:
            type: integer
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseShortConvertDto'
  /api/currency/convert/details:
    get:
      tags:
        - Currency Converter
      summary: >-
        get ammount (in source-currency), source-currency, target-currency and
        convert in to amount in target currency.
      operationId: currencyConvertDetails
      parameters:
        - in: query
          name: source
          schema:
            type: string
        - in: query
          name: target
          schema:
            type: string
        - in: query
          name: amount
          schema:
            type: integer
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CCResponseDto'
components:
  schemas:
    CCRequestDto:
      properties:
        amount:
          type: integer
        sourceCurrency:
          type: string
        targetCurrency:
          type: string
    CCResponseDto:
      properties:
        success:
          type: boolean
        terms:
          type: string
        privacy:
          type: string
        timestamp:
          type: string
        source:
          type: string
        qoutes:
          type: array
          items:
            type: string
    ResponseConvertDto:
      properties:
        query:
          type: object
          $ref: '#/components/schemas/QueryDto'
        results:
          type: object
          $ref: '#/components/schemas/ResultsDto'
        result:
          type: number
          format: double
        createdAt:
          type: string
    ResultsDto:
      type: object
      properties:
        default:
          type: string
      additionalProperties:
        type: object
        $ref: '#/components/schemas/ResultValueDto'
    ResultValueDto:
      type: object
      properties:
        id:
          type: string
        val:
          type: number
          format: double
        to:
          type: string
        fr:
          type: string
    QueryDto:
      properties:
        count:
          type: integer
    ResponseShortConvertDto:
      properties:
        createdAt:
          type: string
        result:
          type: number
          format: double
    ResponseShortVatDetailDto:
      properties:
        countryCode:
          type: string
        validFormat:
          type: boolean
    ResponseVatDetailDto:
      properties:
        company_address:
          type: string
        company_name:
          type: string
        country_code:
          type: string
        query:
          type: string
        success:
          type: boolean
        valid:
          type: boolean
        valid_format:
          type: boolean
        vat_number:  
          type: string